Factored out dag_variable operations.
